+++ spyne/model/_base.py
@@ -325,6 +325,7 @@ class ModelBase(object):
 
         return retval
 
+    # FIXME: Rename this to get_type_name_with_ns_pref
     @classmethod
     def get_type_name_ns(cls, interface):
         """Returns the type name with a namespace prefix, separated by a column.
@@ -567,9 +568,10 @@ class xml:
         information. This is generally a stupid thing to do. Use with caution.
     """
 
-    def __init__(self, root_tag=None, no_ns=False):
+    def __init__(self, root_tag=None, no_ns=False, pretty_print=False):
         self.root_tag = root_tag
         self.no_ns = no_ns
+        self.pretty_print = pretty_print
 
 
 class table: